requirements.txt
Appium-Python-Client==3.2.1
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
cached-property==2.0.1
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
cryptography==44.0.2
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
docutils==0.21.2
et_xmlfile==2.0.0
exchangelib==5.5.1
fire==0.7.0
fonttools==4.56.0
fpdf2==2.7.5
furl==2.1.4
graphviz==0.13.2
h11==0.14.0
holidays==0.45
htmldocx==0.0.6
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
java-access-bridge-wrapper==1.2.0
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kitchen==1.2.6
lxml==5.3.1
more-itertools==10.6.0
mss==6.1.0
natsort==8.4.0
netsuitesdk==1.24.0
notifiers==1.3.3
O365==2.0.26
oauthlib==3.2.2
openpyxl==3.1.5
orderedmultidict==1.0.1
outcome==1.3.0.post0
packaging==24.2
paramiko==3.5.1
pdfminer.six==20221105
pendulum==3.0.0
pillow==10.4.0
platformdirs==4.3.6
ply==3.11
psutil==5.9.8
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pynput-robocorp-fork==5.0.0
pyotp==2.9.0
pypdf==3.17.4
pyperclip==1.9.0
Pypubsub==4.0.3
PySocks==1.7.1
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
pytz==2025.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinauto==0.6.9
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
robocorp-storage==1.0.5
robotframework==7.2.2
robotframework-appiumlibrary==2.1.0
robotframework-archivelibrary==0.4.3
robotframework-assertion-engine==3.0.3
robotframework-databaselibrary==2.1.3
robotframework-datadriver==1.11.2
robotframework-ftplibrary==1.9
robotframework-jsonlibrary==0.5
robotframework-pabot==4.1.1
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.7
robotframework-ride==2.1.2
robotframework-sapguilibrary==1.1
robotframework-seleniumlibrary==6.7.1
robotframework-seleniumtestability==2.1.0
robotframework-sshlibrary==3.8.0
robotframework-stacktrace==0.4.1
rpaframework==30.0.1
rpaframework-core==12.0.1
rpaframework-pdf==8.0.0
rpaframework-windows==9.0.0
rpds-py==0.23.1
scp==0.15.0
selenium==4.15.2
simple-salesforce==1.12.6
six==1.17.0
smartsheet-python-sdk==3.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sspilib==0.2.0
stringcase==1.2.0
tenacity==8.5.0
termcolor==2.5.0
time-machine==2.16.0
trio==0.29.0
trio-websocket==0.12.2
tweepy==3.10.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==4.3.1
uiautomation==2.0.20
urllib3==2.3.0
validators==0.34.0
webdriver-manager==4.0.2
wrapt==1.17.2
wsproto==1.2.0
wxPython==4.2.2
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.3.1
zipp==3.21.0
